Project Detail changes
