+++ src/router/index.js
@@ -3,7 +3,7 @@ import { createRouter, createWebHistory } from 'vue-router'
 import Work from '../views/Work.vue'
 import MakeDecision from '../views/MakeDecision.vue'
 import SelectedProjects from '../views/SelectedProjects.vue'
-import ProjectDetailEdited from '@/views/ProjectDetailEdited.vue'
+import ProjectDetail from '@/views/ProjectDetail.vue'
 // import ProjectDetail from '@/views/ProjectDetail.vue'
 
 
@@ -44,7 +44,7 @@ const router = createRouter({
     {
     path: '/project/:projectSlug', 
     name: 'ProjectDetail',
-    component: ProjectDetailEdited,
+    component: ProjectDetail,
     props: true, // Automatically pass route params as props to the component
   
     },
